requirements.txt
tf-models-official
tensorflow-text
black == 21.6b0
